requirements.txt
torch>=1.9.0
torchvision>=0.10.0
efficientnet-pytorch>=0.7.1
flask>=2.0.0
pillow>=8.0.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
openai>=0.27.0
python-dotenv>=0.19.0
werkzeug>=2.0.0